+++ CHANGELOG.md
@@ -1,9 +1,8 @@
-### 0.3.0 (Next)
+### 0.3.0 (7/19/2015)
 
 * [#5](https://github.com/dblock/slack-ruby-bot/issues/5): Added support for free-formed routes via `match` - [@dblock](https://github.com/dblock).
 * [#6](https://github.com/dblock/slack-ruby-bot/issues/6): Commands and operators take blocks - [@dblock](https://github.com/dblock).
 * [#4](https://github.com/dblock/slack-ruby-bot/issues/4): Messages are posted with `as_user: true` by default - [@dblock](https://github.com/dblock).
-* Your contribution here.
 
 ### 0.2.0 (7/10/2015)